refactor: file allocation
