+++ client/src/context/index.js
@@ -1,3 +1,3 @@
-export { ChatProvider, useChat } from "./ChatProvider/ChatProvider";
-export { RoomProvider, useRoom } from "./RoomProvider/RoomProvider";
-export { UserProvider, useUser } from "./UserProvider/UserProvider";
+export { ChatProvider, useChat } from "./ChatProvider";
+export { RoomProvider, useRoom } from "./RoomProvider";
+export { UserProvider, useUser } from "./UserProvider";
