+++ client/src/reducer/index.js
@@ -1,0 +1,3 @@
+export { chatReducer } from "./chatReducer";
+export { roomReducer } from "./roomReducer";
+export { userReducer } from "./userReducer";
